% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advr.06-package.R
\docType{package}
\name{advr.06-package}
\alias{advr.06}
\alias{advr.06-package}
\title{advr.06: \if{html}{\out{<Title here>}}}
\description{
Implementation of different algorithms to solve the knapsack problem including brute force, dynamic programming, and greedy approaches.
}
\author{
\strong{Maintainer}: Finn Beckmann \email{finbe486@student.liu.se}

Authors:
\itemize{
  \item Max Goldbeck-Löwe \email{maxgo900@student.liu.se}
}

}
\keyword{internal}
